Module Tezos_stdlib_unix__File_event_sink

An implementation of Tezos_base.Internal_event.SINK which writes the events as JSON files in a given directory structure.

It is registered with the URI scheme "unix-files", one can activate it with an URI containing the top-level directory in which the JSON files will be written, e.g. "export TEZOS_EVENTS_CONFIG=unix-files:///the/path/to/write" (the path should be inexistent or already a directory).

The directory structure is as follows: "<section-dirname>/<event-name>/<YYYYMMDD-string>/<HHMMSS-MMMMMM>/<YYYYMMDD-HHMMSS-MMMMMM-xxxx.json>" where "<section-dirname>" is either "no-section" or "section-<section-name>".

type t

module Sink_implementation : Tezos_event_logging.Internal_event.SINK with type t = t

module Query : sig ... end

The module Query provides a fold function over the events stored by a given instantiation of the SINK.t.