Module Tezos_stdlib_unix__Internal_event_unix

Configure the event-logging framework for UNIx-based applications.

module Configuration : sig ... end

val init : ?⁠lwt_log_sink:Tezos_stdlib_unix.Lwt_log_sink_unix.cfg -> ?⁠configuration:Configuration.t -> unit -> unit Lwt.t

Initialize the internal-event sinks by looking at the ?configuration argument and then at the (whitespace separated) list of URIs in the "TEZOS_EVENTS_CONFIG" environment variable, if an URI does not have a scheme it is expected to be a path to a configuration JSON file (cf. Configuration.of_file), e.g.: export TEZOS_EVENTS_CONFIG="unix-files:///tmp/events-unix debug://", or export TEZOS_EVENTS_CONFIG="debug:// /path/to/config.json".

The function also initializes the Lwt_log_sink_unix module (corresponding to the "TEZOS_LOG" environment variable).

val close : unit -> unit Lwt.t

Call close on all the sinks.