Module Tezos_stdlib_unix.Systime_os

val now : unit -> Ptime.t

The current time according to the system clock

val sleep : Ptime.Span.t -> unit Lwt.t

sleep t is an Lwt promise that resolves after t time has elapsed. If t is negative, sleep t is already resolved.